Update instructions for the local storage provisioner manifest file
The /var/lib/rancher/k3s/server/manifests/local-storage.yaml manifest file needs
to be updated on the node to specify the /mnt/data/storage directory for the k3s-1 node.
If this file is not updated then changes to the configmap will be reverted when
the K3s server is restarted.